+++ src/Dropzone/src/Form/DropzoneType.php
@@ -13,6 +13,8 @@ namespace Symfony\UX\Dropzone\Form;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\FileType;
+use Symfony\Component\Form\FormInterface;
+use Symfony\Component\Form\FormView;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
@@ -22,13 +24,21 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
  */
 class DropzoneType extends AbstractType
 {
+    public function buildView(FormView $view, FormInterface $form, array $options): void
+    {
+        $view->vars['multiple'] = $options['multiple'];
+        $view->vars['remove_label'] = $options['remove_label'];
+    }
+
     public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
             'attr' => [
                 'placeholder' => 'Drag and drop or browse',
             ],
+            'remove_label' => 'Remove',
         ]);
+        $resolver->setAllowedTypes('remove_label', 'string');
     }
 
     public function getParent(): string